plantify-docker/mqtt_scripts/requirements.txt
paho-mqtt
mariadb

